requirements.txt
<<<<<<< HEAD
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
numpy>=1.24.0
joblib>=1.3.0
underthesea>=6.8.4
python-dotenv>=1.0.0
=======
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
numpy>=1.24.0
joblib>=1.3.0
underthesea>=6.8.4
python-dotenv>=1.0.0
>>>>>>> fc8bf8d (backend repo)
